java-spring: Adjust API to conform to the specification
Mostly, error responses and HTTP status codes had to be adjusted a bit
on the Backend side.

However, because the Spring framework will validate and reject the
request before the function entry-point on erroneous requests, all we
can expect is a generic error message, where otherwise a more specific
`Argument "foo" is required` would have been expected. For this case,
the test suite had to be slightly adjusted.