Documentation: Improve README about provisioning resources
### Problem
CrateDB needs access to the resource on `COPY ... FROM ...` operations,
but it does not have access even when running within a Docker container
locally, and even less when running in the Cloud.

### Quick solution
The improved documentation section provides an easy way to work around
that limitation by directly accessing the resource(s) on GitHub.